sendAndSave에서 SendMailDto 제거
