+++ mail-server/src/main/kotlin/com/gitub/oopgurus/refactoringproblems/mailserver/Mail.kt
@@ -1,6 +1,5 @@
 package com.gitub.oopgurus.refactoringproblems.mailserver
 
-import com.fasterxml.jackson.databind.ObjectMapper
 import jakarta.mail.internet.MimeMessage
 import org.springframework.mail.javamail.JavaMailSender
 import java.util.concurrent.Executors
@@ -8,58 +7,40 @@ import java.util.concurrent.TimeUnit
 
 class Mail(
         private val javaMailSender: JavaMailSender,
-        private val mailRepository: MailRepository,
-        private val objectMapper: ObjectMapper,
-        private val mailSpamService: MailSpamService
+        private val saveJob: (MailSendStatus) -> Unit
 ) {
 
     private val scheduledExecutorService = Executors.newScheduledThreadPool(10)
 
+    /**
+     * TODO
+     */
     private var afterSeconds: Long? = null
-
     fun afterSeconds(afterSeconds: Long): Mail {
         this.afterSeconds = afterSeconds
         return this
     }
 
-    fun send(mimeMessage: MimeMessage, sendMailDto: SendMailDto) {
-
-        validateToAddress(sendMailDto.toAddress)
-
+    fun send(mimeMessage: MimeMessage) {
         if (afterSeconds != null) {
             val tempSeconds = afterSeconds!!
             scheduledExecutorService.schedule(
-                    { sendMailAndSave(mimeMessage, sendMailDto) },
+                    { sendMailAndSave(mimeMessage) },
                     tempSeconds,
                     TimeUnit.SECONDS
             )
         } else {
-            sendMailAndSave(mimeMessage, sendMailDto)
+            sendMailAndSave(mimeMessage)
         }
     }
 
-    // TODO sendMailDto 제거
-    private fun sendMailAndSave(mimeMessage: MimeMessage, sendMailDto: SendMailDto) {
-        lateinit var mailStatus: MailStatus
-        try {
+    private fun sendMailAndSave(mimeMessage: MimeMessage) {
+        val mailSendSuccess = try {
             javaMailSender.send(mimeMessage)
-            mailStatus = MailSuccessStatus(mailRepository, objectMapper)
+            MailSendStatus.SUCCESS
         } catch (e: Exception) {
-            mailStatus = MailFailureStatus(mailRepository, objectMapper)
+            MailSendStatus.FAILURE
         }
-        mailStatus.save(sendMailDto)
-    }
-
-    private fun validateToAddress(toAddress: String) {
-        mailSpamService.needBlockByDomainName(toAddress).let {
-            if (it) {
-                throw RuntimeException("도메인 차단")
-            }
-        }
-        mailSpamService.needBlockByRecentSuccess(toAddress).let {
-            if (it) {
-                throw RuntimeException("최근 메일 발송 실패로 인한 차단")
-            }
-        }
+        saveJob(mailSendSuccess)
     }
 }
