+++ mail-server/src/main/kotlin/com/gitub/oopgurus/refactoringproblems/mailserver/MailSendStatus.kt
@@ -1,0 +1,6 @@
+package com.gitub.oopgurus.refactoringproblems.mailserver
+
+enum class MailSendStatus(val isSuccess: Boolean) {
+    SUCCESS(true),
+    FAILURE(false),
+}
