--- mail-server/src/main/kotlin/com/gitub/oopgurus/refactoringproblems/mailserver/MailStatus.kt
@@ -1,6 +1,0 @@
-package com.gitub.oopgurus.refactoringproblems.mailserver
-
-interface MailStatus {
-
-    fun save(sendMailDto: SendMailDto): MailEntity
-}
